fall2024-thesis/src/sim/requirements.txt
attrs==24.2.0
cbor==1.0.0
contourpy==1.3.1
coppeliasim_zmqremoteapi_client==2.0.4
cycler==0.12.1
fonttools==4.55.2
kiwisolver==1.4.7
loguru==0.7.3
matplotlib==3.9.3
msgpack==1.1.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
paho-mqtt==2.1.0
pillow==11.0.0
psutil==6.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
scipy==1.14.1
six==1.17.0
